Retry fetching manga entries
This is an experimental retry approach to fetching
from the back-end. There could be some weird hiccups,
stalling request for up to 30 seconds, when a simple
extra request would be done in milliseconds

So now we try this extra request. If it doesn't work,
we'll just show you the page you were still on. This is done
using the re-assignment of pagy object, as we have internal
current page counter, that we need to reset